refactor: delete unused folders
